\name{kplot.mbpls}
\alias{kplot.mbpls}
\alias{kplot.smbpls}
\title{
Multiple plot representation associated with object (s)mbpls
}
\description{
Representation of the scores and weights of the Tables X and Y
}
\usage{
kplot.mbpls(...)
}
\arguments{
  \item{\dots}{
further arguments passed to or from other methods
}
}
\seealso{
}
\examples{
}

