\name{ktabtools-package}
\alias{ktabtools-package}
\alias{ktabtools}
\docType{package}
\title{
Additional R functions related to K-table analyses
}
\description{
This package contains additional R functions related to K-table analyses and Multiblock-(s)pls.
}
\details{
\tabular{ll}{
Package: \tab ktabtools\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-11-21\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Pierre Bady <pierre.bady@unil.ch>
Maintainer: Pierre Bady <pierre.bady@unil.ch>
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
